web-compiler-python/Django/requirements.txt
Django==3.1.6
# psycopg2==2.8
djangorestframework==3.12.4
django-cors-headers==3.7.0